+++ lafleur/mutators/engine.py
@@ -97,6 +97,7 @@ from lafleur.mutators.scenarios_types import (
     SuperResolutionAttacker,
     TypeInstabilityInjector,
     TypeIntrospectionMutator,
+    TypeVersionInvalidator,
 )
 from lafleur.mutators.utils import RedundantStatementSanitizer
 from lafleur.mutators.helper_injection import HelperFunctionInjector
@@ -208,6 +209,7 @@ class ASTMutator:
             MROShuffler,
             BasesRewriteMutator,
             DynamicClassSwapper,
+            TypeVersionInvalidator,
             FrameManipulator,
             ComprehensionBomb,
             SuperResolutionAttacker,